fixed minor grammar issues and typos
I deleted the translation of "for" loop because it doesn't help tos. understand the concept of this kind of loop. Plus, this is a very basic functionality in almost all development languages.
I also replaced the word "extérieurement" by the expression "depuis l'extérieur". It is more common to say it like that in French.